Address review comments on usage_limits PR
Findings from Dirk's review (gist 385eb3425e7bff177a03d0da62498776):

#1 Cold lazy shards: rewrite the godoc rationale in usage_limits.go —
   counting them would NOT force a load (LazyLoadShard.ObjectCountAsync
   reads from disk metadata), but it would do a directory walk + open+read
   per segment metadata file on every write, which is unacceptable on the
   hot path. Document under docs/usage_limits.md "Accepted imperfections"
   with the in-memory-cache extension as a future direction (agreed in
   Slack thread).

#2 Drop WithError (CLAUDE.md rule, Dash0 formatting). Preserve .Warn
   severity — operator action is not required for a transient per-shard
   counting miss.

#3 Tighten shard_write_put.go comment to explicitly call out the
   wire-contract caveat on the forwarded-write path: cluster-API handler
   currently stringifies as HTTP 500 rather than re-typing to 429. Doesn't
   bite the supported deployment shapes (single-node, namespace-pinned
   phase 1), but is now spelled out instead of glossed over.

#4 Rewrite the state.go field godoc: the Manager only governs the
   object-count cap; collections/tenants/shards limits are read at the
   schema-handler use sites, not through the Manager. Drop the inaccurate
   "Set*Counter" wording — only DB.SetUsageLimits exists.

#5 docs/usage_limits.md "CountAsync is O(1)" was wrong; correct to
   O(segments-per-shard) — walks the live segment list summing already-
   loaded counters, no I/O.

#6 Lift gRPC error translation into a shared helper translateTypedError
   and apply it to both makeAuthInterceptor (unary) and
   makeAuthStreamInterceptor (stream). BatchStream now gets the same
   wire contract as the unary surfaces.

#7 docs/usage_limits.md "Accepted imperfections": call out that the
   schema-side caps are not transactional with RAFT — two concurrent
   create requests can both pass and both succeed. Bounded overshoot;
   acceptable at free-tier scale.

#9 Drop cosmetic int64(...) casts in schema/class.go and schema/tenant.go
   — both sides of the comparison are already int.

#10 loopInsertUntilLimit deadline 60s → 120s. The aggressive flush trigger
   is a trigger, not a guarantee; slow CI shouldn't flake this.

#11 (Generated Validate skips enum on empty fields) deferred — fixing it
   requires marking errorCode/limit as required in the OpenAPI spec, which
   forces the generated Go fields from string to *string and would ripple
   through callers. The reviewer hedged ("probably fine"); not worth the
   regen risk this late in the review cycle.

#8 (Manager naming) addressed via the rewritten state.go godoc rather
   than a rename — the type is package-scoped (usagelimits.Manager) and
   the renaming would touch DB-internal Index.usageLimits references for
   little semantic gain. The godoc now makes the single-responsibility
   scope explicit.

Verification: go build, go vet, go test (usagelimits, schema, grpc, rest,
config, objects, db), golangci-lint, goroutine linter, and the full
test/acceptance/usage_limits suite (8 subtests across 2 testcontainers)
all green.

+++ usecases/schema/class.go
@@ -168,7 +168,7 @@ func (h *Handler) AddClass(ctx context.Context, principal *models.Principal,
 		shardCap := dv.Get()
 		if shardCap >= 0 {
 			requested := cls.ShardingConfig.(shardingcfg.Config).DesiredCount
-			if int64(requested) > int64(shardCap) {
+			if requested > shardCap {
 				return nil, 0, usagelimits.NewLimitExceededError(
 					h.errorMessageTemplate(), usagelimits.LimitShards, int64(shardCap))
 			}
